+++ packages/expo-sqlite/dev-plugin-webui/src/app/_layout.tsx
@@ -1,0 +1,13 @@
+import { Slot } from 'expo-router';
+
+import { DevToolsConnectionProvider } from '@/hooks/useDevToolsConnection';
+
+import '../global.css';
+
+export default function RootLayout() {
+  return (
+    <DevToolsConnectionProvider>
+      <Slot />
+    </DevToolsConnectionProvider>
+  );
+}